sound updates

+++ main.py
@@ -311,7 +311,7 @@ class mainPage(QMainWindow):
                 self.connectedCameraOutput_label.clear()
             else:
                 # self.detections[url] = detechYolo.Detech("DetechModel.pt", url, 640, "cpu", "CCTV", self.classes)
-                self.detections.insert(self.activeCam,Camera(url, detechYolo.Detech("DetechModel.pt", url, 640, "cpu", "CCTV",  classes=self.classes, selectedClass=self.selectedClass, user_id=self.user_id), self.cameraWidgets[self.activeCam], str(self.activeCam+1)))
+                self.detections.insert(self.activeCam,Camera(url, detechYolo.Detech("DetechModelS.pt", url, 640, "cpu", "CCTV",  classes=self.classes, selectedClass=self.selectedClass, user_id=self.user_id), self.cameraWidgets[self.activeCam], str(self.activeCam+1)))
                 print("Success!")
                 self.label_27.setStyleSheet("background-color: green")
                 self.label_27.setText("Connected IP: " +str(url))
@@ -701,14 +701,22 @@ class mainPage(QMainWindow):
         )
 
         mycursor = mydb.cursor()
-        query = "DELETE FROM violators"
+        query = f"SELECT * FROM violators WHERE {self.user_id} LIKE user_id"
+        mycursor.execute(query)
+        result = mycursor.fetchall()
+        for row in result:
+            print(f"Removed {row}")
+            os.remove(row[5])
+
+        mycursor = mydb.cursor()
+        query = f"DELETE FROM violators WHERE {self.user_id} LIKE user_id"
         mycursor.execute(query)
         mydb.commit()
 
         #delete in file folder
-        dir = 'violators'
-        for file in os.scandir(dir):
-            os.remove(file.path)
+        # dir = 'violators'
+        # for file in os.scandir(dir):
+        #     os.remove(file.path)
